Upgraded PHP images to Alpine 3.12
So what happened is that on 3.12 the postgresql-dev package depends on
icu-dev, and this causes pecl to build the mongodb extension with ICU support.

This means that now the final image would need the icu-libs package at runtime,
but this would also add a whooping 30MB to the final images.

I could build mongodb manually so that it doesn't pick up the -libicu flag, but
instead I can just use the .so and header files from my own 1maa/postgres:12-src
image.